Add `BranchPythonOperator` to skip over `FileSensor` for old data.
If the data are old then there won't be a `*.done` path indicating the transfer is complete. This adds a `BranchOperator` that will skip the `FileSensor` step if the data are more than a week old. We can adjust this timing on that choice, but this is a proof of concept. 

I thought it better to set it up as a `BranchPythonOperator` so we can visibly see (from the web interface) when an operator was skipped or when it spent time waiting for files.